refactor: separate StoryOperations::Destroy into its own file
